requirements.txt
certifi==2019.3.9
Django==2.1.7
django-cors-headers==2.5.2
django-datetime-utc==1.0.2
django-extensions==2.1.6
djangorestframework==3.9.2
psycopg2==2.8.1
python-dateutil==2.8.0
pytz==2018.9
six==1.12.0
wincertstore==0.2
